+++ logprep/ng/processor/deduplicator/processor.py
@@ -28,7 +28,7 @@ import typing
 from logprep.ng.abc.processor import Processor
 from logprep.processor.base.rule import Rule
 from logprep.processor.deduplicator.rule import DeduplicatorRule
-from logprep.util.helper import add_fields_to, get_dotted_field_value
+from logprep.util.helper import FieldValue, add_fields_to, get_dotted_field_value
 
 
 class Deduplicator(Processor):
@@ -36,7 +36,7 @@ class Deduplicator(Processor):
 
     rule_class = DeduplicatorRule  # type: ignore
 
-    def _apply_rules(self, event: dict, rule: Rule):
+    async def _apply_rules(self, event: dict[str, FieldValue], rule: Rule) -> None:
         rule = typing.cast(DeduplicatorRule, rule)
         for field in rule.fields:
             value_list = get_dotted_field_value(event, field)