+++ logprep/ng/processor/dissector/processor.py
@@ -26,6 +26,7 @@ Processor Configuration
 .. automodule:: logprep.processor.dissector.rule
 """
 
+import typing
 from collections.abc import Callable, Generator
 
 from logprep.ng.processor.field_manager.processor import FieldManager
@@ -45,7 +46,8 @@ class Dissector(FieldManager):
 
     rule_class = DissectorRule  # type: ignore
 
-    def _apply_rules(self, event, rule):
+    async def _apply_rules(self, event: dict[str, FieldValue], rule: Rule) -> None:
+        rule = typing.cast(DissectorRule, rule)
         self.__apply_mapping(event, rule)
         self._apply_convert_datatype(event, rule)